requirements.txt
google-generativeai==0.3.2
pillow==10.2.0
numpy==1.26.4
instagrapi==2.0.3
python-dotenv==1.0.1
noise==1.2.2
colorthief>=0.2.1
webcolors>=1.13.0
apscheduler==3.10.4
requests==2.31.0
opencv-python>=4.8.0
scipy>=1.12.0
pytz==2024.1
resend==0.6.0
fastapi==0.110.0
uvicorn==0.27.1
